Geocoding using the ArcGIS REST API with multifield query

Geocodes addresses given specific address components.This function returns the tibble associated with the query.

For geocoding using a single text string use arc_geo() function.

Usage

arc_geo_multi(
  address = NULL,
  address2 = NULL,
  address3 = NULL,
  neighborhood = NULL,
  city = NULL,
  subregion = NULL,
  region = NULL,
  postal = NULL,
  postalext = NULL,
  countrycode = NULL,
  lat = "lat",
  long = "lon",
  limit = 1,
  full_results = FALSE,
  return_addresses = TRUE,
  verbose = FALSE,
  progressbar = TRUE,
  outsr = NULL,
  langcode = NULL,
  category = NULL,
  custom_query = list()
)

Arguments

address, address2, address3, neighborhood, city, subregion, region, postal, postalext, countrycode

Address components (See Details). At least one field should be different than NA or NULL. A vector of values can be provided for each parameter for multiple geocoding. When using vectors on different parameters, their lengths should be the same.

lat

latitude column name in the output data (default "lat").

long

longitude column name in the output data (default "lon").

limit

maximum number of results to return per input address. Note that each query returns a maximum of 50 results.

full_results

returns all available data from the API service. This is a shorthand of outFields=*. See References. If FALSE (default) only the default values of the API would be returned. See also return_addresses.

return_addresses

return input addresses with results if TRUE.

verbose

if TRUE then detailed logs are output to the console.

progressbar

Logical. If TRUE displays a progress bar to indicate the progress of the function.

outsr

The spatial reference of the x,y coordinates returned by a geocode request. By default is NULL (i.e. the parameter won't be used in the query). See Details.

langcode

Sets the language in which reverse-geocoded addresses are returned.

category

A place or address type that can be used to filter results. Several values can be used as well as a comma-separated string (i.e. "Cinema,Museum). See arc_categories for details.

custom_query

API-specific parameters to be used, passed as a named list.

Value

A tibble with the results. See the details of the output in ArcGIS REST API Service output.

The resulting output would include also the input parameters (columns with prefix q_) for better tracking the results.

Details

More info and valid values in the ArcGIS REST docs

outsr

The spatial reference can be specified as either a well-known ID (WKID). If not specified, the spatial reference of the output locations is the same as that of the service ( WGS84, i.e. WKID = 4326)).

Address components

This function allows to perform structured queries by different components of an address. The following list provides a brief description of each parameter:

address: A string that represents the first line of a street address. In most cases it will be the street name and house number input, but it can also be used to input building name or place-name.

address2: A string that represents the second line of a street address. This can include street name/house number, building name, place-name, or sub unit.

address3: A string that represents the third line of a street address. This can include street name/house number, building name, place-name, or sub unit.

neighborhood: The smallest administrative division associated with an address, typically, a neighborhood or a section of a larger populated place.

city: The next largest administrative division associated with an address, typically, a city or municipality.

subregion: The next largest administrative division associated with an address. Depending on the country, a sub region can represent a county, state, or province.

region: The largest administrative division associated with an address, typically, a state or province.

postal: The standard postal code for an address, typically, a three– to six-digit alphanumeric code.

postalext: A postal code extension, such as the United States Postal Service ZIP+4 code.

countrycode: A value representing the country. Providing this value increases geocoding speed. Acceptable values include the full country name in English or the official language of the country, the two-character country code, or the three-character country code.

References

ArcGIS REST findAddressCandidates

See also

Examples

# \donttest{
simple <- arc_geo_multi(
  address = "Plaza Mayor", limit = 3,
  custom_query = list(outFields = "LongLabel")
)

simple[, c("lat", "lon", "LongLabel")]
#> # A tibble: 3 × 3
#>     lat     lon LongLabel                                    
#>   <dbl>   <dbl> <chr>                                        
#> 1  41.7   -4.73 Plaza Mayor, Valladolid, Castilla y León, ESP
#> 2  39.5   -6.37 Plaza Mayor, Cáceres, Extremadura, ESP       
#> 3  21.2 -102.   Plaza Mayor, León, Guanajuato, MEX           

# Restrict search to Spain
simple2 <- arc_geo_multi(
  address = "Plaza Mayor", countrycode = "ESP",
  limit = 3,
  custom_query = list(outFields = "LongLabel")
)

simple2[, c("lat", "lon", "LongLabel")]
#> # A tibble: 3 × 3
#>     lat   lon LongLabel                                                         
#>   <dbl> <dbl> <chr>                                                             
#> 1  41.7 -4.73 Plaza Mayor, Valladolid, Castilla y León, ESP                     
#> 2  39.5 -6.37 Plaza Mayor, Cáceres, Extremadura, ESP                            
#> 3  41.3 -4.91 Plaza Mayor, 47400, Medina del Campo, Valladolid, Castilla y León…

# Restrict to a region
simple3 <- arc_geo_multi(
  address = "Plaza Mayor", region = "Segovia",
  countrycode = "ESP",
  limit = 3,
  custom_query = list(outFields = "LongLabel")
)

simple3[, c("lat", "lon", "LongLabel")]
#> # A tibble: 3 × 3
#>     lat   lon LongLabel                                                         
#>   <dbl> <dbl> <chr>                                                             
#> 1  41.0 -4.12 Plaza Mayor, 40001, Distrito 1 Seccion 2, Segovia, Castilla y Leó…
#> 2  41.0 -4.12 Plaza Mayor, Calle del Cronista Lecea 2, 40001, Distrito 1 Seccio…
#> 3  40.9 -4.27 Plaza Mayor, 40141, Abades, Segovia, Castilla y León, ESP         
# }