Reverse Geocoding using the ArcGIS REST API

Generates an address from a latitude and longitude. Latitudes must be between [-90, 90] and longitudes between [-180, 180]. This function returns the tibble associated with the query.

Usage

arc_reverse_geo(
  lat,
  long,
  address = "address",
  full_results = FALSE,
  return_coords = TRUE,
  verbose = FALSE,
  progressbar = TRUE,
  custom_query = list()
)

Arguments

lat

latitude values in numeric format. Must be in the range [-90, 90].

long

longitude values in numeric format. Must be in the range [-180, 180].

address

address column name in the output data (default "address").

full_results

returns all available data from the API service. If FALSE (default) only latitude, longitude and address columns are returned.

return_coords

return input coordinates with results if TRUE.

verbose

if TRUE then detailed logs are output to the console.

progressbar

Logical. If TRUE displays a progress bar to indicate the progress of the function.

custom_query

API-specific parameters to be used, passed as a named list (i.e. list(featureTypes = "POI")). See Details.

Value

A tibble with the results.

Details

See https://developers.arcgis.com/rest/geocode/api-reference/geocoding-reverse-geocode.htm for additional parameters to be passed to custom_query.

See also

Examples

# \donttest{

arc_reverse_geo(lat = 40.75728, long = -73.98586)
#> # A tibble: 1 × 3
#>   address                               lat   lon
#>   <chr>                               <dbl> <dbl>
#> 1 W 44th St, New York, NY, 10036, USA  40.8 -74.0

# Several coordinates
arc_reverse_geo(lat = c(40.75728, 55.95335), long = c(-73.98586, -3.188375))
#>   |                                                          |                                                  |   0%  |                                                          |=========================                         |  50%  |                                                          |==================================================| 100%
#> # A tibble: 2 × 3
#>   address                                       lat    lon
#>   <chr>                                       <dbl>  <dbl>
#> 1 W 44th St, New York, NY, 10036, USA          40.8 -74.0 
#> 2 Old Town Chambers, Edinburgh, Scotland, GBR  56.0  -3.19

# With options: zoom to country level
sev <- arc_reverse_geo(
  lat = c(40.75728, 55.95335), long = c(-73.98586, -3.188375),
  custom_query = list(featureTypes = "StreetInt,POI", langCode = "FR"),
  verbose = TRUE, full_results = TRUE
)
#>   |                                                          |                                                  |   0%  |                                                          |=========================                         |  50%  |                                                          |==================================================| 100%

dplyr::glimpse(sev)
#> Rows: 2
#> Columns: 27
#> $ address      <chr> "W 44th St & Broadway, New York, NY, 10036, USA", "Old To…
#> $ lat          <dbl> 40.75732, 55.95325
#> $ lon          <dbl> -73.98596, -3.18827
#> $ Match_addr   <chr> "W 44th St & Broadway, New York, New York, 10036", "Old T…
#> $ LongLabel    <chr> "W 44th St & Broadway, New York, NY, 10036, USA", "Old To…
#> $ ShortLabel   <chr> "W 44th St & Broadway", "Old Town Chambers"
#> $ Addr_type    <chr> "StreetInt", "POI"
#> $ Type         <chr> "", "Hotel"
#> $ PlaceName    <chr> "", "Old Town Chambers"
#> $ AddNum       <chr> "", ""
#> $ Address      <chr> "W 44th St & Broadway", ""
#> $ Block        <chr> "", ""
#> $ Sector       <chr> "", ""
#> $ Neighborhood <chr> "Theater District-Times Square", ""
#> $ District     <chr> "Manhattan", ""
#> $ City         <chr> "New York", ""
#> $ MetroArea    <chr> "New York City Metro", ""
#> $ Subregion    <chr> "New York County", "Edinburgh"
#> $ Region       <chr> "New York", "Scotland"
#> $ RegionAbbr   <chr> "NY", ""
#> $ Territory    <chr> "", ""
#> $ Postal       <chr> "10036", ""
#> $ PostalExt    <chr> "", ""
#> $ CntryName    <chr> "United States", "United Kingdom"
#> $ CountryCode  <chr> "USA", "GBR"
#> $ wkid         <int> 4326, 4326
#> $ latestWkid   <int> 4326, 4326
# }