ESRI (ArcGIS) Spatial Reference data base

Database of available spatial references (CRS) in tibble format.

Format

A tibble with 9,364 rows and fields:

projtype

Projection type ("ProjectedCoordinateSystems", "GeographicCoordinateSystems","VerticalCoordinateSystems")

wkid

Well-Known ID

latestWkid

Most recent wkid, in case that wkid is deprecated

authority

wkid authority (Esri or EPSG)

deprecated

Logical indicating if wkid is deprecated

description

Human-readable description of the wkid

areaname

Use area of the wkid

wkt

Representation of wkid in Well-Known Text (WKT). Useful when working with sf or terra

Source

ESRI Projection Engine factory

Details

This data base is useful when using the outsr parameter of the functions.

Some projections ids have changed over time, for example Web Mercator is wkid = 102100 is deprecated and currently is wkid = 3857. However, both values would work, and they would return similar results.

Note

Data extracted on 14 January 2023.

See also

Examples

# \donttest{
# Get all possible values
data("arc_spatial_references")
arc_spatial_references
#> # A tibble: 9,364 × 8
#>    projtype      wkid latestWkid authority deprecated description areaname wkt  
#>    <chr>        <int>      <int> <chr>     <lgl>      <chr>       <chr>    <chr>
#>  1 ProjectedCo…  2000       2000 EPSG      FALSE      Anguilla 1… Anguill… "PRO…
#>  2 ProjectedCo…  2001       2001 EPSG      FALSE      Antigua 19… Antigua… "PRO…
#>  3 ProjectedCo…  2002       2002 EPSG      FALSE      Dominica 1… Dominic… "PRO…
#>  4 ProjectedCo…  2003       2003 EPSG      FALSE      Grenada 19… Grenada… "PRO…
#>  5 ProjectedCo…  2004       2004 EPSG      FALSE      Montserrat… Montser… "PRO…
#>  6 ProjectedCo…  2005       2005 EPSG      FALSE      St Kitts 1… St Kitt… "PRO…
#>  7 ProjectedCo…  2006       2006 EPSG      FALSE      St Lucia 1… St Luci… "PRO…
#>  8 ProjectedCo…  2007       2007 EPSG      FALSE      St Vincent… St Vinc… "PRO…
#>  9 ProjectedCo…  2008       2008 EPSG      TRUE       NAD 1927 S… Canada … "PRO…
#> 10 ProjectedCo…  2009       2009 EPSG      FALSE      NAD 1927 S… Canada … "PRO…
#> # ℹ 9,354 more rows

# Request with deprecated Web Mercator
library(dplyr)
#> 
#> Attaching package: 'dplyr'
#> The following objects are masked from 'package:stats':
#> 
#>     filter, lag
#> The following objects are masked from 'package:base':
#> 
#>     intersect, setdiff, setequal, union
wkid <- arc_spatial_references %>%
  filter(latestWkid == 3857 & deprecated == TRUE) %>%
  slice(1)

glimpse(wkid)
#> Rows: 1
#> Columns: 8
#> $ projtype    <chr> "ProjectedCoordinateSystems"
#> $ wkid        <int> 102100
#> $ latestWkid  <int> 3857
#> $ authority   <chr> "Esri"
#> $ deprecated  <lgl> TRUE
#> $ description <chr> "WGS 1984 Web Mercator Major Auxiliary Sphere"
#> $ areaname    <chr> "World - 85~S to 85~N"
#> $ wkt         <chr> "PROJCS[\"WGS_1984_Web_Mercator_Auxiliary_Sphere\",GEOGCS[…

add <- arc_geo("London, United Kingdom", outsr = wkid$wkid)

# Note values lat, lon and wkid. latestwkid give the current valid wkid
add %>%
  select(lat, lon, wkid, latestWkid) %>%
  glimpse()
#> Rows: 1
#> Columns: 4
#> $ lat        <dbl> 6711544
#> $ lon        <dbl> -14215.35
#> $ wkid       <int> 102100
#> $ latestWkid <int> 3857
# }